covfefe_api/requirements.txt
numpy
scipy
scikit-learn
torch
tqdm
transformers
nltk
Flask
Google-Search-API
sentence-transformers
